+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 * Update permission name for eholdings module. (UIEH-1104)
 * Provider: Re-order accordions display. (UIEH-1044)
 * Package Record | Change accordion order. (UIEH-1045)
+* Display Holdings Summary Heading for Title record. (UIEH-1109)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)